Updated class TelegramFront.java
