+++ .idea/workspace.xml
@@ -4,13 +4,18 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="a3e43a2d-bbed-4299-b7fb-b77b3b770041" name="Changes" comment="Another small syntax changes">
+    <list default="true" id="a3e43a2d-bbed-4299-b7fb-b77b3b770041" name="Changes" comment="Another small syntax changes, updated class BankUtil.java">
+      <change beforePath="$PROJECT_DIR$/.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/BankUtils/BankUtil.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/BankUtils/BankUtil.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Enums/BankNames.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Enums/BankNames.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Enums/BuySell.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Enums/BuySell.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Rounding/RoundValues.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Rounding/RoundValues.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/TelegramBot/User.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/TelegramBot/User.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/AppLauncher.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/AppLauncher.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/BankUtils/BankUtil.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/bankUtils/BankUtil.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/DailyNotifications/NotificationSender.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/dailyNotifications/NotificationSender.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/Enums/BankNames.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/enums/BankNames.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/Enums/BuySell.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/enums/BuySell.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/PrettyOutput/OutputTextCreator.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/prettyOutput/OutputTextCreator.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/Rounding/RoundValues.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/rounding/RoundValues.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/TelegramBot/TelegramFront.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/telegramBot/TelegramFront.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/TelegramBot/User.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/telegramBot/User.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -225,7 +230,14 @@
       <option name="project" value="LOCAL" />
       <updated>1703516017125</updated>
     </task>
-    <option name="localTasksCounter" value="21" />
+    <task id="LOCAL-00021" summary="Another small syntax changes, updated class BankUtil.java">
+      <created>1703518805881</created>
+      <option name="number" value="00021" />
+      <option name="presentableId" value="LOCAL-00021" />
+      <option name="project" value="LOCAL" />
+      <updated>1703518805881</updated>
+    </task>
+    <option name="localTasksCounter" value="22" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -264,7 +276,7 @@
                   <entry key="branch">
                     <value>
                       <list>
-                        <option value="ComboBoxButtons" />
+                        <option value="TelegramFrontDev" />
                       </list>
                     </value>
                   </entry>
@@ -296,7 +308,8 @@
     <MESSAGE value="Deleted dublicate code in TelegramFront.java" />
     <MESSAGE value="Small syntax changes" />
     <MESSAGE value="Another small syntax changes" />
-    <option name="LAST_COMMIT_MESSAGE" value="Another small syntax changes" />
+    <MESSAGE value="Another small syntax changes, updated class BankUtil.java" />
+    <option name="LAST_COMMIT_MESSAGE" value="Another small syntax changes, updated class BankUtil.java" />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
