+++ src/main/java/AppLauncher.java
@@ -1,4 +1,4 @@
-import TelegramBot.TelegramFront;
+import telegramBot.TelegramFront;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
