+++ src/main/java/enums/BankNames.java
@@ -1,4 +1,4 @@
-package Enums;
+package enums;
 
 public enum BankNames {
     PRIVAT,
