+++ src/main/java/enums/BuySell.java
@@ -1,4 +1,4 @@
-package Enums;
+package enums;
 
 public enum BuySell {
     BUY,
