+++ src/main/java/rounding/RoundValues.java
@@ -1,4 +1,4 @@
-package Rounding;
+package rounding;
 
 import java.text.DecimalFormat;
